pyproject.toml
[tool.poetry]
name = "api-with-db"
version = "0.1.0"
description = ""
authors = ["jekxkan <smirnova.eyu@mail.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = '^3.11'
fastapi = '^0.115.0'
ruff = '^0.9.2'
sqlalchemy = '^2.0.37'
pydantic-settings = '^2.7.1'
alembic = '^1.14.1'
asyncpg = '^0.30.0'
uvicorn = '^0.34.0'
psycopg2-binary = "^2.9.10"


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
